+++ CHANGELOG.md
@@ -1,9 +1,14 @@
+## 5.1.3
+
+- Added support for range inputs that display their value as you slide it around.
+- Updated dependencies.
+
 ## 5.1.2
 
 - Added a search icon to search inputs.
 - Added support for select elements with the multiple attribute.
 - Added styling to tables within a form, or tables with the `semanticForm` class.
-- Fixed missing styles to single radios.
+- Fixed missing styles on single radios.
 - Fixed a bug that caused additional text labels to not work on checkbox groups, single checkboxes, and radios.
 - Fixed alignment issues on checkbox and radio groups.
 - Updated dependencies.